backend/requirements.txt
fastapi==0.58.1
uvicorn==0.11.5
pydantic-sqlalchemy==0.0.4
